+++ packages/sharing-editor/public/samples/011_component_gallery/pages/status.status.py
@@ -1,12 +1,12 @@
-import time
+import asyncio
 import streamlit as st
 
 with st.status("Downloading data..."):
     st.write("Searching for data...")
-    time.sleep(2)
+    await asyncio.sleep(2)
     st.write("Found URL.")
-    time.sleep(1)
+    await asyncio.sleep(1)
     st.write("Downloading data...")
-    time.sleep(1)
+    await asyncio.sleep(1)
 
 st.button("Rerun")