+++ packages/sharing-editor/public/samples/011_component_gallery/pages/status.status1.py
@@ -1,13 +1,13 @@
-import time
+import asyncio
 import streamlit as st
 
 with st.status("Downloading data...", expanded=True) as status:
     st.write("Searching for data...")
-    time.sleep(2)
+    await asyncio.sleep(2)
     st.write("Found URL.")
-    time.sleep(1)
+    await asyncio.sleep(1)
     st.write("Downloading data...")
-    time.sleep(1)
+    await asyncio.sleep(1)
     status.update(label="Download complete!", state="complete", expanded=False)
 
 st.button("Rerun")